Do more work in the hook
